renamed 'useData.js' -> 'dataUtils.js'. added handleRemoveSelectedTag() in utils to use axios to communicate with backend.
