+++ frontend/src/components/AdminPage/AdminPage.jsx
@@ -3,7 +3,10 @@ import { Row, Col } from "react-bootstrap"; // Make sure to import Bootstrap com
 import { ButtonList } from "./ButtonList";
 import { ProgramInfo } from "./ProgramInfo";
 import { TagInfo } from "./TagInfo";
-import useFetchData from "./useData";
+import useFetchData, {
+  handleRemoveSelectedTag,
+  useDeleteData,
+} from "./dataUtils";
 import { useAuth0 } from "@auth0/auth0-react"; // Import the Auth0 hook
 import styled from "styled-components";
 
@@ -180,12 +183,13 @@ function AdminPage() {
   };
 
   const handleRemoveTag = () => {
-    // remove the selected tag (if there is one)
-    if (selectedTag != null) {
-      console.log("request remove tag: " + selectedTag.toString());
-      // give a warning?
-      // make call to backend to remove this tag (and associated programTag entries) from the database
-    }
+    const onSuccess = (data) => {
+      // Handle successful deletion
+    };
+    const onError = (error) => {
+      // Handle error
+    };
+    handleRemoveSelectedTag(selectedTag, onError, onSuccess);
   };
 
   useEffect(() => {
